% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_helper_functions.R
\name{filter_dt}
\alias{filter_dt}
\title{filter dt according filter in pt_options}
\usage{
filter_dt(x, pt_options)
}
\arguments{
\item{x}{data.table containing data}

\item{pt_options}{pt_options list}
}
\value{
data.table containing grouped/pivoted data
}
\description{
filter dt according filter in pt_options
}

