% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_helper_functions.R
\name{group_dt}
\alias{group_dt}
\title{group a data.table according to rows/columns of pt_options}
\usage{
group_dt(x, pt_options, pt_group)
}
\arguments{
\item{x}{data.table containing data}

\item{pt_options}{pt_options list}

\item{pt_group}{vector of field names of x to be used in by statement of data.table}
}
\value{
data.table containing grouped data
}
\description{
group a data.table according to rows/columns of pt_options
}

