% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_helper_functions.R
\name{group_pivot_dt}
\alias{group_pivot_dt}
\title{group then pivot dt according to row/cols of pt_options}
\usage{
group_pivot_dt(x, pt_options)
}
\arguments{
\item{x}{data.table containing data}

\item{pt_options}{pt_options list}
}
\value{
data.table containing grouped/pivoted data
}
\description{
group then pivot dt according to row/cols of pt_options
}

