% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_helper_functions.R
\name{group_row_col_dt}
\alias{group_row_col_dt}
\title{group a data.table according to rows/columns of pt_options}
\usage{
group_row_col_dt(x, pt_options)
}
\arguments{
\item{x}{data.table containing data}

\item{pt_options}{pt_options list}
}
\value{
data.table containing subtotals based on row/cols in pt_options
}
\description{
group a data.table according to rows/columns of pt_options
}

