% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_helper_functions.R
\name{new_pt_options}
\alias{new_pt_options}
\title{pt_options hold options passed to pivot table functions}
\usage{
new_pt_options(pt_filter = NA_character_, pt_row = NA_character_,
  pt_col = NA_character_, pt_metrics = NA_character_,
  pt_sort = NA_character_, pt_sort_order = 1, pt_row_subtotals = FALSE,
  pt_col_subtotals = FALSE, pt_grand_totals = NA_character_,
  pt_other = NULL)
}
\arguments{
\item{pt_filter}{character that is parsed in context of dt i}

\item{pt_row}{character vector - rows to group by}

\item{pt_col}{character vector - columns to pivot and group by}

\item{pt_metrics}{character that is parsed in context of dt j}

\item{pt_sort}{character vecotr used to sort dt}

\item{pt_sort_order}{1 decreasing (default) or -1 increasing. can be vector.}

\item{pt_row_subtotals}{logical - should row subtotals be included?}

\item{pt_col_subtotals}{logical - should column subtotals be included?}

\item{pt_other}{to be implemented}
}
\description{
pt_options hold options passed to pivot table functions
}

