% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_helper_functions.R
\name{pivot_dt}
\alias{pivot_dt}
\title{pivot a data.table according to rows/columns of pt_options}
\usage{
pivot_dt(x, pt_options)
}
\arguments{
\item{x}{data.table containing data}

\item{pt_options}{pt_options list}
}
\value{
data.table containing pivoted data
}
\description{
pivot a data.table according to rows/columns of pt_options
}

