% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_helper_functions.R
\name{pt_calc}
\alias{pt_calc}
\title{pt_options hold options passed to pivot table functions}
\usage{
pt_calc(x, pt_options)
}
\arguments{
\item{x}{data.table containing data}

\item{pt_options}{pt_options list}
}
\value{
data.table altered according to pt_options
}
\description{
pt_options hold options passed to pivot table functions
}

