% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_helper_functions.R
\name{sort_dt}
\alias{sort_dt}
\title{sort dt according sort in pt_options}
\usage{
sort_dt(x, pt_options)
}
\arguments{
\item{x}{data.table containing data}

\item{pt_options}{pt_options list}
}
\value{
data.table containing grouped/pivoted data
}
\description{
sort dt according sort in pt_options
}

