% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_helper_functions.R
\name{subtotal_dt}
\alias{subtotal_dt}
\title{add subtotals to a dt using pt_options rows/cols/subtotals}
\usage{
subtotal_dt(x, pt_options)
}
\arguments{
\item{x}{data.table containing data}

\item{pt_options}{pt_options list}
}
\value{
data.table containing subtotals based on row/cols in pt_options
}
\description{
add subtotals to a dt using pt_options rows/cols/subtotals
}

